Simplified vector subtraction
Instead of subtracting the coordinates respectively, I took advantage of already-existent operator+ and negative sign, making subtraction only a negative addition. I did not thought this tweak would make a real difference, but it turned out that the runtime of problem 8-3 on PDOGS is reduced from 16ms to 7ms, which is actually quite a leap.